thêm API chức năng hiển thị danh sách lịch sử thi của 1 bài thi

+++ src/main/java/app/dto/history/LastPlayedResponse.java
@@ -14,6 +14,6 @@ public class LastPlayedResponse {
     private long correct;
     private long wrong;
     private long timeTaken;
-    private long score;
+    private double score;
     private List<QuestionResultResponse> questions;
 }
